Backend tests and frontend first steps

+++ services/src/main/kotlin/pt/isel/markdown2slides/ProjectInfoService.kt
@@ -27,6 +27,8 @@ class ProjectInfoService(
         ownerId: UUID,
         projectId: UUID,
     ): Either<ProjectError, Unit> = trxManager.run {
+        val project = repoProjectInfo.findById(projectId) ?: return@run failure(ProjectError.ProjectNotFound)
+        if (project.ownerId != ownerId) return@run failure(ProjectError.ProjectNotAuthorized)
         repoProjectInfo.deleteById(projectId)
         val deleted = repoProjectContent.deleteProjectContent(ownerId, projectId)
         if(deleted) success(Unit) else failure(ProjectError.ContentDeletionError)
@@ -40,19 +42,24 @@ class ProjectInfoService(
     }
 
     fun getProjectDetails(
+        ownerId: UUID,
         projectId: UUID
     ): Either<ProjectError, ProjectInfo> = trxManager.run {
         val project = repoProjectInfo.findById(projectId) ?: return@run failure(ProjectError.ProjectNotFound)
+        if(project.ownerId != ownerId) return@run failure(ProjectError.ProjectNotAuthorized)
         success(project)
     }
 
     fun editProjectDetails(
+        ownerId: UUID,
         projectId: UUID,
         name: String?,
         description: String?,
         visibility: Visibility?
     ): Either<ProjectError, ProjectInfo> = trxManager.run {
         val project = repoProjectInfo.findById(projectId) ?: return@run failure(ProjectError.ProjectNotFound)
+        if(project.ownerId != ownerId) return@run failure(ProjectError.ProjectNotAuthorized)
+
 
         val updated = project.copy(
             name = name ?: project.name,
@@ -69,9 +76,12 @@ class ProjectInfoService(
 }
 
 sealed class ProjectError {
-    data object SomeError : ProjectError()
     data object ContentDeletionError : ProjectError()
     data object ProjectNotFound : ProjectError()
+    data object ProjectNotAuthorized : ProjectError()
     data object ImageDeletionError : ProjectError()
     data object ImageNotFound : ProjectError()
+    data class DatabaseError(val message: String) : ProjectError()
+    data class FileSystemError(val message: String) : ProjectError()
+    data class UnknownError(val message: String) : ProjectError()
 }
